docs(createDataGrid): split the examples into composable + sub-component files
Decompose the editing, pinned, and spanning examples so no file runs much past ~150 lines: logic moves to use*.ts composables, peripheral UI (toolbar/footer/history) to small sub-components, and the table + grid wiring stays in each entry. Also folds in the in-progress no-op-edit guard (click-away commits only when the value changed) and the spanning Today-column pin, fixes a post-merge createTimeline<TimelineTicket<EditEntry>> typing, and corrects top-level-ref template bindings (:stats / :summary, not .value).
